+++ webui/src/test/setup.ts
@@ -1,0 +1,12 @@
+import '@testing-library/jest-dom/vitest'
+import { cleanup } from '@testing-library/react'
+import { afterEach } from 'vitest'
+import i18n from '../i18n'
+
+void i18n.changeLanguage('en')
+
+afterEach(() => cleanup())
+
+if (!Element.prototype.scrollIntoView) {
+  Element.prototype.scrollIntoView = () => undefined
+}